receipt_analyzer/requirements.txt
Flask
boto3
Werkzeug
requests
python-dotenv
Pillow